Relocate BookInfo
