+++ src/components/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import Navbar from './Navbar';
-import BookInfo from './BookInfo';
+import BookInfo from '../containers/BookInfo';
 import BookCatalog from '../containers/BookCatalog';
 import About from './About';
